4th Commit

+++ spring_rest_api_pro1/src/main/java/com/spring_rest_api_pro1/spring_rest_api_pro1/entities/User.java
@@ -31,7 +31,7 @@ public class User {
     @Column(name = "last_name")
     private String lastName;
 
-    private int contact;
+    private Long contact;
 
     private String address;
